relocate surgical masks & stringtable
