+++ @twc_framework/addons/twc_medical_surgery/stringtable.xml
@@ -1,0 +1,20 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project name="TWC Medical Surgery">
+	<Package name="TWC_Medical_Surgery">
+		<Key ID="STR_TWC_SURGERY_HANDED">
+			<Original>%1 handed over the patient to a surgical team</Original>
+			<English>%1 handed over the patient to a surgical team</English>
+			<German>%1 handed over the patient to a surgical team</German>
+		</Key>
+		<Key ID="STR_TWC_SURGERY_COMPLETED">
+			<Original>%1 was released by the surgical team</Original>
+			<English>%1 was released by the surgical team</English>
+			<German>%1 was released by the surgical team</German>
+		</Key>
+		<Key ID="STR_TWC_FIELDSURGERY_COMPLETED">
+			<Original>%1 was fully cared for by %2</Original>
+			<English>%1 was fully cared for by %2</English>
+			<German>%1 was fully cared for by %2</German>
+		</Key>
+	</Package>
+</Project>
