+++ @twc_framework/addons/twc_medical_surgery/ui/surgical_mask.paa
